added dataAccess.js module for the database
