+++ scripts/database.js
@@ -5,7 +5,7 @@
     modules to get copies of the state.
 
 */
-const database = {
+export const database = {
     orderBuilder: {},
     styles: [
         { id: 1, style: "Classic", price: 500 },
@@ -42,70 +42,70 @@ const database = {
     ]
 }
 
-export const addCustomOrder = () => {
-    // Copy the current state of user choices
-    const newOrder = {...database.orderBuilder}
+// export const addCustomOrder = () => {
+//     // Copy the current state of user choices
+//     const newOrder = {...database.orderBuilder}
 
-    const totalCustomOrders = database.customOrders.length
+//     const totalCustomOrders = database.customOrders.length
 
-    // database.orders.length > 0 ?
+//     // database.orders.length > 0 ?
     
     
-    if (totalCustomOrders < 1) {
-        newOrder.id = 1
-    } else {newOrder.id = [...database.customOrders].pop().id + 1}
+//     if (totalCustomOrders < 1) {
+//         newOrder.id = 1
+//     } else {newOrder.id = [...database.customOrders].pop().id + 1}
   
-    // newOrder.id = [...database.customOrders].pop().id + 1
+//     // newOrder.id = [...database.customOrders].pop().id + 1
 
-    // Add a timestamp to the order
-    newOrder.timestamp = Date.now()
+//     // Add a timestamp to the order
+//     newOrder.timestamp = Date.now()
 
-    // Add the new order object to custom orders state
-    database.customOrders.push(newOrder)
+//     // Add the new order object to custom orders state
+//     database.customOrders.push(newOrder)
 
-    // Reset the temporary state for user choices
-    database.orderBuilder = {}
+//     // Reset the temporary state for user choices
+//     database.orderBuilder = {}
 
-    // Broadcast a notification that permanent state has changed
-    document.dispatchEvent(new CustomEvent("stateChanged"))
-}
+//     // Broadcast a notification that permanent state has changed
+//     document.dispatchEvent(new CustomEvent("stateChanged"))
+// }
 
-export const getCurrentOrder = () => {
-    return {...database.orderBuilder}
-}
+// export const getCurrentOrder = () => {
+//     return {...database.orderBuilder}
+// }
 
-export const getMetals = () => {
-    return [...database.metals]
-}
+// export const getMetals = () => {
+//     return [...database.metals]
+// }
 
-export const getStyles = () => {
-    return [...database.styles]
-}
+// export const getStyles = () => {
+//     return [...database.styles]
+// }
 
-export const getSizes = () => {
-    return [...database.sizes]
-}
+// export const getSizes = () => {
+//     return [...database.sizes]
+// }
 
-export const getOrders = () => {
-    return [...database.customOrders]
-}
+// export const getOrders = () => {
+//     return [...database.customOrders]
+// }
 
-export const getJewelryTypes = () => {
-    return [...database.jewelryTypes]
-}
+// export const getJewelryTypes = () => {
+//     return [...database.jewelryTypes]
+// }
 
-export const setJewelryType = (id) => {
-    database.orderBuilder.jewelryTypeId = id
-}
+// export const setJewelryType = (id) => {
+//     database.orderBuilder.jewelryTypeId = id
+// }
 
-export const setMetal = (id) => {
-    database.orderBuilder.metalId = id
-}
+// export const setMetal = (id) => {
+//     database.orderBuilder.metalId = id
+// }
 
-export const setSize = (id) => {
-    database.orderBuilder.sizeId = id
-}
+// export const setSize = (id) => {
+//     database.orderBuilder.sizeId = id
+// }
 
-export const setStyle = (id) => {
-    database.orderBuilder.styleId = id
-}
+// export const setStyle = (id) => {
+//     database.orderBuilder.styleId = id
+// }
